GVT-3297: Rename search-utils.tsx to search-dropdown-item.tsx
