+++ ui/src/asset-search/search-dropdown-item.tsx
